Simulation procedure
Just download the 6 files, start MATLAB and enter 'gui_simulate'.
You will find a Mpar.mat file after filling the table and clicking
'Start Simulation' in the current folder (type 'pwd' in the command
window to get the current folder)
The structure of the parameter matrix is the same as in the gui table.
Further infos 'help startSimulation'